tswRev: center window & fix sleep
- Place TSW game window to a better location on the screen (see Rev11 in tswRev.asm)
- Change the sleep duration between movement tweens to more reasonable values (see comment below)
- Fix a small bug when displaying the title of messagebox (see common.rb)
- Add more detailed usage for tswRev.rb

The default Windows sleep/timer resolution is 1/64 second (https://stackoverflow.com/questions/21156944/how-to-get-the-current-windows-system-wide-timer-resolution),
and if the sleep duration is not integer multiple of the resolution, the real sleep time can be either n or n-1 times the resolution (https://learn.microsoft.com/en-us/windows/win32/api/synchapi/nf-synchapi-sleep#remarks).
Therefore, the sleep duration set prior to this commit is not reasonable. However, if tswBGM is running, i.e., the TSW game thread is playing multimedia,
the resolution will become 1 ms at the cost of more energy consumption (https://stackoverflow.com/questions/74975185/how-to-wait-for-a-millisecond),
which can be verified by the first link. So, 1 ms sleep duration, under this case, will be real 1 ms, which is too short so the animation does not look smooth.
Therefore, the current sleep interval is more close to integer multiple of 15.6 ms, so under both cases, there won't be a significant discrepancy.